+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.2 (November 23, 2022)
 - Add the cmdlet `Get-SecurityTxtFile`.
 - Fixed the user agent so that it shows the current version.
+- Removed all references to "security.txt" being a draft standard, because it's not anymore;  it's RFC 9116!
 
 ## Version 1.1 (October 25, 2022)
 - Added ShouldProcess support for `New-SecurityTxtFile`. (PSScriptAnalyzer)