+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # SecurityTxtToolkit change log
 
-## Version 1.2 (November 23, 2022)
+## Version 1.3.0
+- Add support for verifying digitally-signed "security.txt" files, if GnuPG is available.
+
+## Version 1.2.0 (November 23, 2022)
 - Add the cmdlet `Get-SecurityTxtFile`.
 - Fixed the user agent so that it shows the current version.
 - Removed all references to "security.txt" being a draft standard, because it's not anymore;  it's RFC 9116!
 
-## Version 1.1 (October 25, 2022)
+## Version 1.1.0 (October 25, 2022)
 - Added ShouldProcess support for `New-SecurityTxtFile`. (PSScriptAnalyzer)
 - Refactored `Test-SecurityTxtFile` cmdlet.